committing work on feedbacks on B2B

+++ lib/core/connect_end/model/get_incoming_order_ddetail_response_model/item.dart
@@ -14,7 +14,7 @@ class Item {
   String? imageUrl;
   int? quantity;
   double? unitPrice;
-  int? lineTotal;
+  double? lineTotal;
   int? minimumOrderQuantity;
   int? packSize;
   String? unit;
