+++ lib/core/connect_end/model/list_incoming_orders_response_model/order.dart
@@ -15,7 +15,7 @@ class Order {
   String? customerName;
   int? itemsOrdered;
   List<Item>? items;
-  int? orderTotal;
+  double? orderTotal;
   String? currency;
   DateTime? dateOrdered;
   String? paymentStatus;
